Implemented some architectures
Implemented UNet3+ achitecture (rather crude with lots of hardcoded stuff).
Training of UNet3+ takes a lot of VRAM and compute! Also it is very dependent on initialization (see loss plots where the initialized Encoder showed much better behaviour on a mock dataset.
Even plain UNet, which needs a lot less compute and VRAM, shows promising results.
The different sandbox*.py files are intended for trying out different things, keep in mind that for visualization you have to specify the right architecture to be able to load the weights accordingly.